requirements.txt
# LIGHTHOUSE Practice Intelligence Platform - Dependencies

# Core
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI / LLM
anthropic>=0.39.0

# Encryption
cryptography>=41.0.0
argon2-cffi>=23.1.0

# Document ingestion
pdfplumber>=0.10.0
unstructured>=0.16.0
python-pptx>=0.6.21
python-docx>=1.1.0

# Web dashboard
streamlit>=1.32.0
